move ModuleMCP to misc
